requirements.txt
boto3==1.28.22
botocore==1.31.22
cffi==1.15.1
cryptography==41.0.3
javaobj-py3==0.4.3
jmespath==1.0.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodomex==3.18.0
pyjks==20.0.0
pyOpenSSL==23.2.0
python-dateutil==2.8.2
s3transfer==0.6.1
six==1.16.0
twofish==0.3.0
urllib3==1.26.16
